Fixed migrations and added uploads scaffolding

+++ db/migrate/20170707061220_create_devices.rb
@@ -3,8 +3,8 @@ class CreateDevices < ActiveRecord::Migration[5.1]
     create_table :devices do |t|
       t.string :uniqueIdentifier
       t.string :type
-      t.double :latitude
-      t.double :longitude
+      t.decimal :latitude
+      t.decimal :longitude
 
       t.timestamps
     end
